+++ src/index.ts
@@ -4,8 +4,8 @@ import { PromisePool } from './promise-pool'
 
 export default PromisePool
 
+export * from './contracts'
 export * from './promise-pool'
 export * from './promise-pool-error'
 export * from './return-value'
-export * from './stoppable'
 export * from './stop-the-promise-pool-error'